[cuda] read 32-bit discretized histogram as int64 in the best-split finder
When a leaf's max per-bin stat (num_data_in_leaf * num_grad_quant_bins) reaches
65536, the discretized histogram uses 32-bit bins: the construct kernel stores
each bin as an int64 (grad32 << 32 | hess32) via int64 atomics, and the leaf's
packed total is likewise int64. But the best-split finder dispatched the 32-bit
(non-16-bit) case with BIN_HIST_TYPE = int32_t and read the histogram through an
`int32_t*` offset by task->hist_offset. That reads 4-byte half-bins at the wrong
stride (alternating grad/hess halves of consecutive int64 bins), so the finder
saw garbage for every leaf that needed 32-bit bins -- i.e. any leaf with enough
data. The result was near-random splits (root split gain ~0.2 vs the correct
~25000) and useless models whenever a dataset was large enough to require 32-bit
histograms (with num_grad_quant_bins=16 this is any leaf with >= 4096 rows).

The 8-bit and 16-bit paths were correct (and were the only paths exercised by the
small-data parity tests), which is why this stayed hidden until large-scale
quantized training. Fix: read the 32-bit histogram as int64 (BIN_HIST_TYPE =
int64_t, pointer cast to const int64_t*, offset in bin units), in both the
forward and reverse split-search branches.

Before -> after (CUDA quantized vs CPU quantized, num_grad_quant_bins=16):
  n=2000  (16-bit leaves): corr 1.0 (already correct)
  n=8000  (32-bit root):   corr -0.004 -> corr 1.0, maxΔ 0
  n=200000:                corr  0.005 -> corr 1.0, maxΔ 0
Quantized model quality now matches double-precision (RMSE within 0.01%), and
large-scale quantized training is usable (and modestly faster) on CUDA.

+++ src/treelearner/cuda/cuda_best_split_finder.cu
@@ -976,9 +976,13 @@ __global__ void FindBestSplitsDiscretizedForLeafKernel(
             // output parameters
             out);
         } else {
-          const int32_t* hist_ptr =
-            reinterpret_cast<const int32_t*>(IS_LARGER ? larger_leaf_splits->hist_in_leaf : smaller_leaf_splits->hist_in_leaf) + task->hist_offset;
-          FindBestSplitsDiscretizedForLeafKernelInner<USE_RAND, USE_L1, USE_SMOOTHING, false, int32_t, int64_t, false, false>(
+          // 32-bit histogram: each bin is an int64 (grad32 << 32 | hess32), written by
+          // the construct kernel via int64 atomics. It must be read as int64 (offset in
+          // bin units). Reading it as int32 read 4-byte half-bins at the wrong stride and
+          // produced garbage splits whenever a leaf needed 32-bit bins (large leaves).
+          const int64_t* hist_ptr =
+            reinterpret_cast<const int64_t*>(IS_LARGER ? larger_leaf_splits->hist_in_leaf : smaller_leaf_splits->hist_in_leaf) + task->hist_offset;
+          FindBestSplitsDiscretizedForLeafKernelInner<USE_RAND, USE_L1, USE_SMOOTHING, false, int64_t, int64_t, false, false>(
             // input feature information
             hist_ptr,
             // input task information
@@ -1030,9 +1034,10 @@ __global__ void FindBestSplitsDiscretizedForLeafKernel(
             // output parameters
             out);
         } else {
-          const int32_t* hist_ptr =
-            reinterpret_cast<const int32_t*>(IS_LARGER ? larger_leaf_splits->hist_in_leaf : smaller_leaf_splits->hist_in_leaf) + task->hist_offset;
-          FindBestSplitsDiscretizedForLeafKernelInner<USE_RAND, USE_L1, USE_SMOOTHING, true, int32_t, int64_t, false, false>(
+          // 32-bit histogram is int64-per-bin; read as int64 (see forward branch above).
+          const int64_t* hist_ptr =
+            reinterpret_cast<const int64_t*>(IS_LARGER ? larger_leaf_splits->hist_in_leaf : smaller_leaf_splits->hist_in_leaf) + task->hist_offset;
+          FindBestSplitsDiscretizedForLeafKernelInner<USE_RAND, USE_L1, USE_SMOOTHING, true, int64_t, int64_t, false, false>(
             // input feature information
             hist_ptr,
             // input task information
